moved ccfri store out of the facility store
